requirements-dev.txt
selenium==3.141.0
coverage
codecov
django-stubs
flake8
